Restructure api-pls-express-router package
