+++ packages/api-pls-express-router/middleware/not-allowed.js
@@ -1,8 +1,8 @@
 'use strict';
 
-const serverErrors = require('../../util/server-errors');
-const sendJson = require('../../util/send-json');
-const log = require('../../util/log');
+const serverErrors = require('../util/server-errors');
+const sendJson = require('../util/send-json');
+const log = require('../util/log');
 
 module.exports = function(req, res) {
   log.info({req}, 'An action that is not allowed was attempted at an endpoint.');
